マップのテキスト部分を変れるようにした。あとは画像
